Server and client files for assignment 6
Both files work. They work when both are being run from the localhost,
and when both are being run on the uhunix, but they don't work if one is
one my home comuter and the other is on the uhunix. I think UH most
likely blocks external clients from just connecting to any port.